requirements.txt
gradio>=4.0.0
openai-whisper>=20231017
langchain>=0.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langdetect>=1.0.9
torch>=2.0.0
numpy>=1.24.0
accelerate>=0.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain-community>=0.4.0
google-generativeai>=0.8.0